+++ src/main/java/warmup/one/WarmupOneTenthSet.java
@@ -18,4 +18,16 @@ public class WarmupOneTenthSet {
         }
         return counter >= 1 && counter <= 3;
     }
+
+    /**
+     * Given two non-negative int values, return true if they have the same last digit, such as
+     * with 27 and 57. Note that the % "mod" operator computes remainders, so 17 % 10 is 7.
+     *
+     * @param a
+     * @param b
+     * @return
+     */
+    public boolean lastDigit(int a, int b) {
+        return b % 10 == a % 10;
+    }
 }